testy dla cpu

+++ src/cpuclassifier/CPUCalculateClassP.hpp
@@ -75,16 +75,17 @@ namespace nbc4cpu {
     }
     Record computing;
     // 1. Calculate exponent
-    for (int i = 0; i < record.size(); i++)
+    for (long unsigned int i = 0; i < record.size(); i++)
       computing.emplace_back(record[i] - avgs[i]);
     // Note : x - avg
+
     for (auto& x : computing) x = -1 * (x * x);
     // Note : -(x - avg)^2
 
-    for (int i = 0; i < computing.size(); i++)
+    for (long unsigned int i = 0; i < computing.size(); i++)
       computing[i] = exp(computing[i] / (2.0 * variances[i]));
 
-    for (int i = 0; i < computing.size(); i++)
+    for (long unsigned int i = 0; i < computing.size(); i++)
       computing[i] = computing[i] / (sqrt(variances[i]) * sqr2pi);
 
     // 3. summarize
